mediawiki: Enable a subset of default extensions
Full list of default extensions in MediaWiki
https://www.mediawiki.org/wiki/Bundled_extensions_and_skins

This initial set of extensions are chosen from:
- extensions that I've used myself over the past 3 years
- testing done in #1267

Extensions that are advanced features for administrators, meant for spam
control, advanced tags or suitable only to a specific kind of user are
not included.

Fixes #1382

More extensions can be enabled in the future if they're generally useful.

Signed-off-by: Joseph Nuthalapati <njoseph@riseup.net>
Reviewed-by: Sunil Mohan Adapa <sunil@medhas.org>