help: Fix failing setup when manual directory is not available
- It is reported on the mailing list that first setup operation on the help app
fails. The logs indicate that /usr/share/freedombox/manual directory is not
present on the system. Although this situation does not occur on a normal
installation, catch, log and ignore this error to be safe.

- Also ensure that this static files component is added in post_init so that
basic setup is not at all affected by it.

Tests:

- In the development directory, move doc/manual to doc/manual.bak. Observe that
the error message is printed during initialization but the process continues.

- With the manual directory properly restored, the full help manual is displayed
properly with images.

Signed-off-by: Sunil Mohan Adapa <sunil@medhas.org>
Reviewed-by: James Valleroy <jvalleroy@mailbox.org>